$(feat) : add go mod instead of vendor issue #7
