+++ examples/async/go.mod
@@ -1,0 +1,3 @@
+module async
+
+go 1.13
