+++ examples/nomal/go.mod
@@ -1,0 +1,5 @@
+module normal
+
+go 1.13
+
+require github.com/jhunters/bigqueue v1.0.1 // indirect
