+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/jhunters/bigqueue
+
+go 1.13
